# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for user management",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/projects": {
            "get": {
                "description": "Retrieve all projects from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Projects.Project"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new project in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "parameters": [
                    {
                        "description": "Project object",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Projects.Project"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID of the created project",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all projects from MongoDB",
                "tags": [
                    "Projects"
                ],
                "summary": "Delete all projects",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/projects/search": {
            "get": {
                "description": "Search projects in MongoDB based on a keyword",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Search projects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Projects.Project"
                            }
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "description": "Retrieve a project by its ID from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get a project by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Projects.Project"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing project in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated project object",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Projects.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Delete a project by its ID from MongoDB",
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Projects.Project": {
            "type": "object",
            "required": [
                "pl_appstore",
                "pl_body",
                "pl_cli",
                "pl_doc",
                "pl_embedded",
                "pl_github",
                "pl_googleplay",
                "pl_image",
                "pl_linux",
                "pl_macos",
                "pl_package",
                "pl_title",
                "pl_windows"
            ],
            "properties": {
                "pl_appstore": {
                    "type": "string",
                    "example": "http://example.com/appstore"
                },
                "pl_body": {
                    "type": "string",
                    "example": "Description of my project"
                },
                "pl_cli": {
                    "type": "string",
                    "example": "http://example.com/cli"
                },
                "pl_doc": {
                    "type": "string",
                    "example": "http://example.com/doc"
                },
                "pl_embedded": {
                    "type": "string",
                    "example": "http://example.com/embedded"
                },
                "pl_github": {
                    "type": "string",
                    "example": "http://example.com/github"
                },
                "pl_googleplay": {
                    "type": "string",
                    "example": "http://example.com/googleplay"
                },
                "pl_id": {
                    "type": "string"
                },
                "pl_image": {
                    "type": "string",
                    "example": "http://example.com/image.jpg"
                },
                "pl_linux": {
                    "type": "string",
                    "example": "http://example.com/linux"
                },
                "pl_macos": {
                    "type": "string",
                    "example": "http://example.com/macos"
                },
                "pl_package": {
                    "type": "string",
                    "example": "http://example.com/package"
                },
                "pl_title": {
                    "type": "string",
                    "example": "My Project"
                },
                "pl_windows": {
                    "type": "string",
                    "example": "http://example.com/windows"
                }
            }
        },
...
    }
}